Restructure the buffer module
Moves `Pixel` into the trait module.

Exposes a public `buffer` module where iterators and other auxiliary
definitions relating to the `ImageBuffer` are exported.
